stm32: split interface from implementation of DMA wrapper
Idea is to have common interface and different implementation for
different stm32 MCU families.
